Create new controller
Create new overground walking controller to stimulate H-Reflex every ten strides (so long as within the walkway bounds, that is, not during a turn) that incorporates the gait event detection state machine present in the other treadmill controllers but uses kinematics (difference in z-axis component of the ankle markers) to detect the swing phase (rather than stance phase) of each leg (Note that the threshold was determined and set based on a couple of participants from the C3 study and may not be optimal). See 'TODO' notes for areas of future improvement or optimization. NWB