Add stride-by-stride H-reflex stimulation
Update 'NirsHreflexOpenLoopWithAudio' controller to include the option to flexibly stimulate the H-reflex on a stride-by-stride basis by handling optional input arrays 'stimL' and 'stimR' (of the same format as 'velL' and 'velR'). Update the protocol script to stimulate more frequently during the first and last 'Practice' blocks than the middle ones. Add the 'stimL' and 'stimR' arrays to the Spinal Adaptation study profiles in the 'GenerateProfileSpinalStudy' function.