Update C3 protocol script
Update 'RunProtocol_C3' to include the force deletion flag to ensure that profiles that are generated that should not be used are permanently deleted. Fix bugs with incorrectly spelled call to 'fullfile' and converting the 10-Meter Walk Test times from char arrays to double. Add handling for case when profiles have already been generated whether in session 1 or 2. NWB