Rename Arduino code
Rename Arduino code file to use the gait event state machine to determine the timing of stimulation in a treadmill speed independent manner. Add code block to receive input as to whether to stimulate in the current stride from the MATLAB controller. Update stimulation conditions to remove the step count condition since whether to stimulate on the current stride is being communicated by the MATLAB controller. NWB