Remove directory changing code
Remove code in 'generateProfiles_C3' function that retrieves the working directory changes the directory and then changes back to the working directory at the end to replace it with code to simply save the profile using the full file path. Update comments. Remove the conditional statement for the speed profiles and instead use the minimum of two possible fast speeds before dividing by two for the slow speed to ensure the 2:1 ratio. Add structure to define the number of strides in short trials and long trials. Remove unnecessary variable clearing. These changes are a subset of suggestions from ChatGPT to simplify this function. NWB