Add booleans
Add two boolean variables to store whether the left or right leg should be stimulated on the current stride based on input from the MATLAB controller, which has access to the stimulation profile (which sets the stimulation on a stride-by-stride basis). Remove the unnecessary 'numStrides' variable since stimulation will be based on input from the MATLAB controller. Comment out the imposed delay, which is handled by the conditional statement. NWB