Fix bug in data transfer function
Fix bug in function to automatically transfer all of the Spinal Adaptation study data by adding the '...' to ensure proper cell array concatenation. Fix array indexing and dimensionality of the 'startsGaps' and 'endsGaps' arrays that was causing a bug in array assignment. Add 'TODO' note to fix bug in the case of no gaps for a particular marker. NWB