Fix bug with trial indexing
Fix bug where trial number was being incorrectly used as the trial index in the array leading to the wrong trial being processed. Add check to ensure that no trial indices passed as input that are not present in the data folder are mistakenly processed. Remove irrelevant call to non-existent 'Disconnect' method of Vicon Nexus SDK. Handle case of no gaps for a particular marker. NWB