Remove 'delay' call
Remove call to the Arduino 'delay' function, and replace it with code to check the elapsed time since the 'delay' function prevents any other operations from running and may be leading to missed stimuli when running the Arduino version of the controller. Update comments. Add set up code to specify the mode of the output pins. Add 'Serial.println()' commands to help with debugging. NWB