Add separate 'canStim' variables
Add separate variables for whether the system 'canStim' for the right and left leg to prevent stimulation of one leg being blocked by the initiation of stimulation on the other leg. Delete unused and inaccessible 'Serial.println' commands. Simplify 'if' statement in main loop to use 'else' instead of two 'if's. Remove unused 'else' statement. Update comments. Uncomment Serial Monitor output that will hopefully be useful for troubleshooting the next iteration of the code. NWB