Change mid-single stance estimation
Update the estimation of 50% of the single stance gait phase to use the learning factor continuously updating the single stance duration estimate on a stride-by-stride basis for a better head-to-head comparison (i.e., more fair) with the Arduino gait event state machine version of the code. NOTE that these changes will need to be reverted if wanting to replicate the procedure used during early stages of spinal adaptation study H-reflex testing. Make minor formatting change to the Arduino code for the gait event state machine. NWB