Update C3 study protocols for controls
Update the C3 study protocol scripts to handle the case of control participants for which we need to know the participant with stroke for which the participant is serving as a match to load the correct speed profiles. Print the speeds for controls, too, to confirm correct profiles used. Update comments for clarity. Remove generation of 'SyncFiles' folder from 'transferData' function since now automatically generated by C3D2MAT. NWB